\name{defineEvaluationModes}
\alias{defineEvaluationModes}
\title{
Define evaluation modes}
\description{
Get all predefined evaluation mode names
}
\usage{
defineEvaluationModes()
}
\value{
A \emph{\code{vector}} of strings, each representing a reusable evaluation mode name.
}
\author{
\packageAuthor{wyz.code.offensiveProgramming}

Maintainer: \packageMaintainer{wyz.code.offensiveProgramming}
}
\seealso{
Refer to \bold{\emph{\code{\link{EvaluationMode}}}}.
}
\examples{
##---- typical case ----
defineEvaluationModes()
}
\keyword{ programming }
\keyword{ utilities }
\concept{ code evaluation mode }

