\name{defineFunctionReturnTypesParameterName}
\alias{defineFunctionReturnTypesParameterName}
\title{
define function return type parameter name
}
\description{
Provides the parameter name to use to define function return type.
}
\usage{
defineFunctionReturnTypesParameterName()
}\author{
\packageAuthor{wyz.code.offensiveProgramming}

Maintainer: \packageMaintainer{wyz.code.offensiveProgramming}
}
\seealso{
Refer to \code{\link{EvaluationMode}}.
}
\examples{
##---- typical case ----
defineFunctionReturnTypesParameterName()
}
\keyword{ programming }
\keyword{ utilities }
\concept{ code evaluation mode }
