\name{defineTestCaseDefinitionsParameterName}
\alias{defineTestCaseDefinitionsParameterName}
\title{
Test case definition parameter name}
\description{
Define the paramater name to hold test case definitions
}
\usage{
defineTestCaseDefinitionsParameterName()
}
\value{
A single string that is the parameter name to use.
}
\author{
\packageAuthor{wyz.code.offensiveProgramming}

Maintainer: \packageMaintainer{wyz.code.offensiveProgramming}
}
\seealso{
Refer to \code{\link{EvaluationMode}}. See sibling \code{\link{EvaluationMode}}.
}
\examples{
##---- typical case ----
defineTestCaseDefinitionsParameterName()
}
\keyword{ programming }
\keyword{ utilities }
\concept{ code evaluation mode }
