\name{getEllipsisName}
\alias{getEllipsisName}
\title{Get ellipsis.}
\description{
Get ellipsis argument name value.
}
\usage{
getEllipsisName()
}
\value{
A \emph{\code{string}} with value \bold{\code{"\ldots"}}, no more no less.
}
\author{
\packageAuthor{wyz.code.offensiveProgramming}

Maintainer: \packageMaintainer{wyz.code.offensiveProgramming}
}
\examples{
# typical test
getEllipsisName()
#[1] "..."
}
