\name{isAuditable}
\alias{isAuditable}
\title{Is Auditable}
\description{ Retrieve option telling if code is auditable }
\usage{ isAuditable() }
\value{
A \emph{\code{boolean}} value. To turn value to \emph{\code{TRUE}},  set option
\bold{\emph{\code{op_audit}}} to \emph{\code{TRUE}}.
}
\author{
\packageAuthor{wyz.code.offensiveProgramming}

Maintainer: \packageMaintainer{wyz.code.offensiveProgramming}
}
\examples{
# typical case
isAuditable()
# FALSE
}
