\name{getEllipsisName}
\alias{getEllipsisName}
\title{Get ellipsis.}
\description{
Use method \code{getEllipsisName} to get ellipsis argument value.
}
\usage{
getEllipsisName()
}
\value{
A string with value \code{\ldots}, no more no less.
}
\author{
\packageAuthor{wyz.code.offensiveProgramming}

Maintainer: \packageMaintainer{wyz.code.offensiveProgramming}
}
\examples{
# typical test
getEllipsisName()
#[1] "..."
}
