\name{isAuditable}
\alias{isAuditable}
\title{Is auditable.}
\description{
Use method \code{isAuditable} to know if code is auditable
}
\usage{
isAuditable()
}
\value{
A boolean value. To turn value to \code{TRUE},  set option
\code{op_audit} to \code{TRUE}.
}
\author{
\packageAuthor{wyz.code.offensiveProgramming}

Maintainer: \packageMaintainer{wyz.code.offensiveProgramming}
}
\examples{
# typical case
isAuditable()
# FALSE
}
